+++ src/noise_suppression/__init__.py
@@ -1,5 +1,5 @@
 from noisereduce import reduce_noise as suppress_by_noise_sample
 
-from src.noise_suppression.model.nn_noise_suppressor import NeuralNetworkNoiseSuppressor
+from src.noise_suppression.nn.nn_noise_suppressor import NeuralNetworkNoiseSuppressor
 
 __all__ = ['suppress_by_noise_sample', 'NeuralNetworkNoiseSuppressor']
