+++ src/noise_suppression/nn/nn_noise_suppressor.py
@@ -5,7 +5,7 @@ import numpy as np
 import torch
 from scipy.signal import convolve
 
-from src.noise_suppression.model._demucs import Demucs
+from src.noise_suppression.nn._demucs import Demucs
 
 _DEMUCS_CFG = {
     'chin': 1,
